backend/api/requirements.txt
flask
flask-cors
gunicorn
pillow
pytesseract
opencv-python-headless
requests
PyMuPDF
pandas
python-dotenv
google-genai
groq
PyJWT
werkzeug
